Refactor Makefiles/folders for DSA & Fix gitignore
Switched LSP from ccls to clangd and wanted to create standardized ways
for me to create compile_commands.json w/ bear. Switched folder
structure to typical C library with lib and include.

Small refactor for hashtable to accept void * instead of HashObject, to
standardize for whatever I plan to store in the table. In this case it's
for the JSON parser I am working on.