Create handlers for each route
The idea is that when you have routes you will create
a Requesthandler struct, pass it the function that
builds the response with what you want to execute if
the path fits, and then add your request handlers to
an array in a global wrangler that will be passed to
the request handler function.

The c file abstracts the header building and path finding
for anyone creating a web server.